+++ src/Home-Login/Login.js
@@ -1,5 +1,7 @@
-import React, { Component, useState } from 'react'
-import './auth.css';
+import React, {useState} from "react"
+import useGlobalState from "../state";
+import {Link} from "react-router-dom";
+import "./auth.css";
 
 const styles = {
     backgroundColor: "black",
@@ -9,25 +11,71 @@ const styles = {
     flexDirection: "column",
     alignItems: "center",
     justifyContent: "space-between",
-
 }
- class Register extends Component {
 
-    render() {
+export default ({history}) => {
 
-      return (
+    const [email, setEmail] = useState("")
+    const [password, setPassword] = useState("")
+
+    // global state
+    const [, setToken] = useGlobalState("token")
+    const [, setUser] = useGlobalState("user")
+
+    const handleEmail = e => setEmail(e.target.value)
+    // console.log(handleEmail)
+
+    const handlePassword = e => setPassword(e.target.value)
+    // console.log(handlePassword)
+
+    const handleSubmit = async (e) => {
+
+        try {
+            const url = "http://localhost:4000/users/login"
+            // stop form reloading aka browser default behavior
+            e.preventDefault()
+
+            // fetch request (backend)
+            const results = await fetch(url, {
+                // fetch options
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify({email, password}),
+            })
+
+            // parse the token
+            const fetchData = await results.json();
+
+            // higher order function
+            // take the initial value of token and modify it
+            setToken(() => fetchData.token)
+            // console.log(fetchData.token)
+
+            setUser(() => fetchData.user)
+            // console.log(fetchData.user)
+
+            history.push("/")
+
+        } catch (err) {
+            alert(err.message)
+        }
+    }
+
+    return (
         <div style={styles}>
             <div className='register'>
-            <form className="register-form">
-              <input type="text" placeholder="username"/>
-              <input type="password" placeholder="password"/>
-              <button>login</button>
-              <p >Not registered? <a href="#">Create an account</a></p>
-            </form>
-               </div>
-</div>
-        )
-    }
+                <form className="register-form" onSubmit={handleSubmit}>
+                    <label htmlFor='email'>Email</label>
+                    <input type="text" id='email' autoComplete="off" onChange={handleEmail}/>
+                    <label htmlFor='password'>Password</label>
+                    <input autoComplete="off" type="password" id='password' value={password} onChange={handlePassword}/>
+                    <button type='submit'>Login</button>
+                    <p className="message">Not registered? <Link to="/register">Create an account</Link></p>
+                </form>
+            </div>
+        </div>
+    )
 }
 
-export default Register
