+++ src/components/SideBar/SideBar.js
@@ -1,5 +1,6 @@
 import React, {useState, useEffect} from 'react'
 import { Link } from 'react-router-dom'
+import SetCountries from '../SetCountries/SetCountries'
 import classes from "./SideBar.module.css"
 import Top3 from "./TOP3/Top3"
 
@@ -9,11 +10,11 @@ import Top3 from "./TOP3/Top3"
      const info = props.covid.filter(i => {
         return i.cases > 1000000
      }).map(i => {
-         return (<Top3 
-         country={i.country} 
-         cases={i.cases} 
+         return (<Top3
+         country={i.country}
+         cases={i.cases}
          deaths={i.deaths}
-         recovered={i.recovered} 
+         recovered={i.recovered}
          />)
      })
 
@@ -30,13 +31,13 @@ import Top3 from "./TOP3/Top3"
         countriesList()
     }, [])
     const isShown = () => {
-        
+
     }
-    
-    
-    
+
+
+
     const country = props.countryObject.country ? (<h3>{props.countryObject.country}</h3>) : "Please Search Your Country"
-    
+
 
     return (
         <div className={classes.SideBar}>
@@ -68,13 +69,8 @@ import Top3 from "./TOP3/Top3"
                 </div>
             </div>
             <div className={classes.Section3}>
-            <h1>Track Your Countries</h1>
-                <div className={classes.Section3Title}>
-                    <p>Country</p>
-                    <p>Total Cases</p>
-                    <p className={classes.D}>Deaths</p>
-                    <p className={classes.R}>Recovered</p>
-                </div>
+                <h1>Track Your Countries</h1>
+                <SetCountries/>
             </div>
             <Link className={classes.Link} to='/auth'>Please Sign Up To Add a Country</Link>
         </div>
