+++ src/config/routes.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {Switch, Route, BrowserRouter} from 'react-router-dom'
 import Login from '../Home-Login/Login'
-import SignUp from '../Home-Login/SignUp'
+import Register from '../Home-Login/Register'
 import Home from '../Pages/Home'
 import Profile from '../Pages/Profile'
 
@@ -11,7 +11,7 @@ export default (props) => (
             {/* <Route exact path='/' component={ Homepage } /> */}
             <Route path={'/'} exact component={Home}/>
             <Route path='/login' component={Login}/>
-            <Route path='/signup' component={SignUp}/>
+            <Route path='/register' component={Register}/>
             <Route path={'/profile'} component={Profile}/>
         </Switch>
     </BrowserRouter>
